Rejigger how we launch processes
One strategy for 1.9, 1.8, and JRuby.

Not using ChildProcess at this point, because it's not clear
how it provides access to all the output, is poorly documented
and adds a layer of abstraction that seems to be not working
out for aruba at all.  I don't want that at this time.