Refactored a few things to avoid unsafe metaprogramming
Closes #24

Basically, automatically include CLILogging with Main is bad, in that
you can't control the addition of the log convenience methods apart
from use of Main.

Further, including Main and CLILogging into the Object space in the bin
file as the ONLY way of doing things is less than ideal.  This is still
possible, but now you can wrap your app inside a class, and have
the same client interface, but with all metaprogammatically-added
things scoped to that class.