Subtle usability enhancement when both devices are TTYs
If you set the level of a CLILogger to something higher than WARN,
previously, this would have no effect on the error logger.
Generally, this is OK, but when both devices are TTYs, it makes it
appear as if the logger is ignoring the log level.

So, in that case, we set the log level of the error logger as well.